+++ src/app/code/page.tsx
@@ -3,7 +3,6 @@ import { redirect } from "next/navigation";
 
 import { auth } from "@/actions/auth";
 import logo from "@/assets/logo.svg";
-import { Header } from "@/components/Header";
 import { LoginForm } from "@/components/login-form";
 import { NormalLayout } from "@/components/ui/layout/NormalLayout";
 
@@ -15,9 +14,7 @@ export default async function CodePage() {
   }
 
   return (
-    <NormalLayout>
-      <Header title="Login" />
-
+    <NormalLayout title="Login">
       <div className="relative h-[180px] w-full">
         <Image
           src={logo}
