+++ src/services/team.ts
@@ -8,3 +8,9 @@ export const getTeamByCode = async (code: string) => {
     where: eq(teamsTable.code, code),
   });
 };
+
+export const getTeamById = async (id: string) => {
+  return await db.query.teamsTable.findFirst({
+    where: eq(teamsTable.id, id),
+  });
+};
